Splits deploy logic for connections and rules in different files
