+++ server/lib/auth0/index.js
@@ -1,0 +1,13 @@
+import * as rules from './rules'
+import * as connections from './connections'
+
+export default {
+  /* COnnection and database operations */
+  validateDatabases : connections.validateDatabases,
+  updateDatabases : connections.updateDatabases,
+
+  /* rules operations */
+  validateRules : rules.validateRules,
+  deleteRules : rules.deleteRules,
+  updateRules : rules.updateRules,
+}
